refactor: logo adjustments
