--- src/components/logo.js
@@ -1,11 +1,0 @@
-import { Typography } from '@mui/material';
-
-const Logo = ({ className }) => {
-  return (
-    <Typography component="h1" variant="h3" className={className}>
-      Billy
-    </Typography>
-  );
-};
-
-export default Logo;
